+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "autisticstuff"
-version = "0.1.2b2"
+version = "0.1.3a1"
 description = "Autistic stuff, either created by us, or stolen from someone on the internet"
 readme = "README.md"
 authors = [
@@ -9,16 +9,14 @@ authors = [
 ]
 requires-python = ">=3.13"
 dependencies = [
+    "aiogram>=3.22.0",
+    "aiohttp>=3.12.15",
+    "fastapi>=0.116.1",
     "loguru>=0.7.3",
     "pydantic>=2.11.7",
     "python-dotenv>=1.1.1",
     "sqlalchemy[asyncio]>=2.0.43",
-]
-
-[project.optional-dependencies]
-dev = [ "psycopg2[binary]>=2.9.10", "ruff>=0.12.10" ]
-
-    
+]   
 
 [build-system]
 requires = ["uv_build>=0.8.13,<0.9.0"]
