+++ src/autisticstuff/__init__.py
@@ -1,9 +1,8 @@
-from . import (
-	astuff_logs as astuff_logs,
-	astuff_progbars as astuff_progbars,
-	astuff_sqla as astuff_sqla,
-	astuff_utils as astuff_utils,
-)
-
-
-__all__ = ["astuff_logs", "astuff_progbars", "astuff_sa", "astuff_utils"]
+"""
+autisticstuff
+-> fastapi: respsys (unified response system and exc handler, APIExceptions and more), utils
+-> logs: logging system with interception
+-> safe: RetryableClientSession(aiohttp.ClientSession), safe aiogram funcs
+-> sqla: layers (BaseMapping, BaseRepository), types (PydanticJSON), utils (update_with_callback, validate_kwargs_for_model)
+-> utils: config (AppSettings, SettingsField), enums (for URS), lru (lru-cache wrappers for objects), progbars
+"""
