--- src/autisticstuff/astuff_progbars/__init__.py
@@ -1,3 +1,0 @@
-from .enums import Styles
-
-__all__ = [Styles]
