--- src/autisticstuff/astuff_sqla/layers/mapping/__init__.py
@@ -1,4 +1,0 @@
-from .inst import get_base_mapping
-
-
-__all__ = ["get_base_mapping"]
