--- src/autisticstuff/astuff_sqla/layers/repository/__init__.py
@@ -1,4 +1,0 @@
-from .inst import _BaseRepository
-
-
-__all__ = ["_BaseRepository"]
