--- src/autisticstuff/astuff_utils/configuration/__init__.py
@@ -1,4 +1,0 @@
-from .inst import AppSettings, SettingsField
-
-
-__all__ = ["AppSettings", "SettingsField"]
