+++ src/autisticstuff/fastapi/respsys/__init__.py
@@ -1,0 +1,6 @@
+"""Unified Response System"""
+
+from .handler import apply_exception_handler, exception_handler
+
+
+__all__ = ["apply_exception_handler", "exception_handler", "exceptions", "schemas"]
