+++ src/autisticstuff/sqla/layers/__init__.py
@@ -1,0 +1,5 @@
+from .mapping import get_base_mapping
+from .repo import _BaseRepository
+
+
+__all__ = ["_BaseRepository", "get_base_mapping"]
