+++ src/autisticstuff/utils/_dev/placeholders.py
@@ -1,0 +1,2 @@
+def _raise_ni_error():
+	raise NotImplementedError
