+++ src/autisticstuff/utils/config/__init__.py
@@ -1,0 +1,4 @@
+from .config import AppSettings, SettingsField
+
+
+__all__ = ["AppSettings", "SettingsField"]
