+++ src/autisticstuff/utils/enums/__init__.py
@@ -1,0 +1,4 @@
+from .enums import APIErrorSpecs, StatusCodeMap
+
+
+__all__ = ["APIErrorSpecs", "StatusCodeMap"]
